requirements.txt
azure-cognitiveservices-speech==1.24.0
Flask==2.2.2
PyYAML==6.0
requests==2.28.2
wechatpy==1.8.18
